poseonly_8obj_nosubject_seeblack

+++ training_scripts/train.py
@@ -56,9 +56,9 @@ DEBUG = False
 BS = 4  
 # SAVE_STEPS = [500, 1000, 2000, 5000, 10000, 15000, 20000, 25000, 30000] 
 # VLOG_STEPS = [4, 50, 100, 200, 500, 1000]   
-VLOG_STEPS = [1000, 5000, 10000, 50000, 70000, 80000, 90000, 100000] 
+VLOG_STEPS = [100, 10000, 50000, 100000, 150000, 200000]  
 # SAVE_STEPS = copy.deepcopy(VLOG_STEPS) 
-SAVE_STEPS = [10000, 20000, 30000, 40000, 50000, 60000, 70000, 80000, 90000, 100000]  
+SAVE_STEPS = [100, 500, 1000, 5000, 10000, 20000, 30000, 40000, 50000, 60000, 70000, 80000, 90000, 100000, 110000, 120000, 130000, 140000, 150000, 160000, 170000, 180000, 190000, 200000]   
 NUM_SAMPLES = 18     
 NUM_COLS = 4   
 
@@ -366,7 +366,7 @@ def infer(args, step_number, wandb_log_data, accelerator, unet, scheduler, vae, 
             "bnha cat", 
             "bnha elephant", 
             "bnha bus", 
-            "bnha giraffe", 
+            # "bnha giraffe", 
             "bnha jeep", 
         ] 
 
@@ -1188,7 +1188,7 @@ def parse_args(input_args=None):
     parser.add_argument(
         "--stage2_steps",
         type=int,
-        default=100000,
+        default=200000,
         help="Number of steps for stage 2 training", 
     )
     parser.add_argument(
@@ -1743,7 +1743,7 @@ def main(args):
     input_embeddings = torch.clone(accelerator.unwrap_model(text_encoder).get_input_embeddings().weight).detach()  
 
     # steps_per_angle = {} 
-    if DEBUG: 
+    if DEBUG and accelerator.is_main_process: 
         if osp.exists(f"vis"): 
             shutil.rmtree(f"vis") 
         os.makedirs("vis")  
@@ -1782,7 +1782,7 @@ def main(args):
         # Convert images to latent space
         vae.to(accelerator.device, dtype=weight_dtype)
 
-        if DEBUG: 
+        if DEBUG and accelerator.is_main_process: 
             for batch_idx, img_t in enumerate(batch["pixel_values"]): 
                 img = (img_t * 0.5 + 0.5) * 255  
                 img = img.permute(1, 2, 0).cpu().numpy().astype(np.uint8) 
@@ -2352,6 +2352,8 @@ def main(args):
                 wandb_log_data = infer(args, step, wandb_log_data, accelerator, unet, noise_scheduler, vae, text_encoder, continuous_word_model, merger) 
                 force_wandb_log = True 
                 set_seed(args.seed + accelerator.process_index) 
+
+            torch.cuda.empty_cache() 
 
             if DEBUG: 
                 for p_, p in zip(unet_params_safe, unet.parameters()):
